requirements.txt
fastapi
requests
jinja2
uvicorn
python-chess
python-multipart


